Extracted TranslatorImpl to its own file
